soap-student-service/requirements.txt
spyne
sqlalchemy
psycopg2-binary
lxml
requests